+++ config/install/field.field.node.fits_technical_metadata.field_data_of.yml
@@ -1,0 +1,20 @@
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.node.field_data_of
+    - node.type.fits_technical_metadata
+id: node.fits_technical_metadata.field_data_of
+field_name: field_data_of
+entity_type: node
+bundle: fits_technical_metadata
+label: 'Data of'
+description: ''
+required: false
+translatable: false
+default_value: {  }
+default_value_callback: ''
+settings:
+  handler: 'default:node'
+  handler_settings: {  }
+field_type: entity_reference
