+++ config/install/ield.field.media.fits_technical_metadata.field_media_file.yml
@@ -1,0 +1,26 @@
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.media.field_media_file
+    - media.type.fits_technical_metadata
+  module:
+    - file
+id: media.fits_technical_metadata.field_media_file
+field_name: field_media_file
+entity_type: media
+bundle: fits_technical_metadata
+label: File
+description: ''
+required: true
+translatable: true
+default_value: {  }
+default_value_callback: ''
+settings:
+  file_directory: '[date:custom:Y]-[date:custom:m]'
+  file_extensions: xml
+  max_filesize: ''
+  description_field: false
+  handler: 'default:file'
+  handler_settings: {  }
+field_type: file
