+++ config/install/node.type.fits_technical_metadata.yml
@@ -1,0 +1,17 @@
+langcode: en
+status: true
+dependencies:
+  module:
+    - menu_ui
+third_party_settings:
+  menu_ui:
+    available_menus:
+      - main
+    parent: 'main:'
+name: 'Fits Technical Metadata'
+type: fits_technical_metadata
+description: 'Convenience node to hold indexable FITS values'
+help: ''
+new_revision: true
+preview_mode: 1
+display_submitted: true
